requirements.txt
crispy-bootstrap5==2024.10
Django==5.1.3
django-allauth==65.2.0
django-browser-reload==1.17.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-json-widget==2.0.1
django-jsonform==2.23.1
django-mysql==4.15.0
django-npm-mjs==3.1.0
django-simple-history==3.7.0
gunicorn==23.0.0
JSON-minify==0.3.0
mysqlclient==2.2.6
pre-commit==4.0.1
python-dotenv==1.0.1
selenium==4.27.1
webdriver-manager==4.0.2
